loyal/serializers/customer_serializer.py
from rest_framework import serializers
from loyal.models import Customer

class CustomerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Customer
        fields = ('first_name', 'last_name', 'email')
